+++ PollGroup/EPoll/Architectures/Linux_x64.cs
@@ -11,7 +11,7 @@ internal sealed partial class Linux_x64 : IArch<epoll_event_packed>
     public static partial int epoll_close(int epfd);
 
     [LibraryImport("libc", SetLastError = true)]
-    public static partial int epoll_ctl(int epfd, epoll_op op, int fd, ref epoll_event_packed ee);
+    public static partial int epoll_ctl(int epfd, epoll_op op, nint fd, ref epoll_event_packed ee);
 
     [LibraryImport("libc", SetLastError = true)]
     public static partial int epoll_wait(int epfd, epoll_event_packed[] ee, int maxevents, int timeout);
